Move restaurant typings and data fetching
