+++ src/types/restaurant.d.ts
@@ -1,0 +1,10 @@
+interface Restaurant {
+  id: string;
+  name: string;
+  location: string;
+  rating: number;
+  tags: string[];
+  description: string;
+  images: string[];
+  docId: string;
+}
